+++ NEWS.md
@@ -1,4 +1,9 @@
-# SpaceMarkers development version
+# SpaceMarkers 1.1.2
+
+* Added getPairwiseInteractingGenes which enables pairwise analysis of 
+interacting patterns 
+
+# SpaceMarkers 1.1.1
 
 * `getSpatialFeatures`: add default method to infer the object passed to it.